Fix initial render of "post like" icon via mount hack
We need to wait for the component to mount on the client before rendering the likes. That's because we can only determine whether the post is liked by a user when we have the auth cookie at hand.
Without this check, the server will render the "unliked" heart, because he never has the user ID. After hydration at the client, the unliked heart is not replaced with the liked heart, because no properties (e.g. noOfLikes) have changed (because we provide the data via hydration/React Query), so even though a user liked a post, the unliked heart will be shown.