perf: Remove Prisma client from client bundle
For the each enum model, Prisma does not only give you a `const` from `@prisma/client` but also the type as string literal. This commit removes all usage of the `const` in favor of string literals - that way the Prisma client is not part of the client bundle.
Read more: https://github.com/prisma/prisma/issues/13567#issuecomment-1332030096