feat: Scroll to top of page on navigation
This is a hack to automatically scroll to the top of a page when the user navigates to another route.
Usually this is done via `Link`'s `scroll` attribute. In our case, it is not the page who is scrollable, but only the container containing the page content, so we have to do it manually.
See: https://github.com/vercel/next.js/discussions/45715